pyproject.toml
[project]
name = "ai-for-engineering-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "datasets>=3.5.0",
    "debugpy>=1.8.14",
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.14",
    "langchain-postgres>=0.0.13",
    "numpy>=2.2.4",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "pgvector>=0.3.6",
    "pip>=25.0.1",
    "psycopg>=3.2.6",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pyhumps>=3.8.0",
    "python-jose>=3.4.0",
    "qdrant-client>=1.13.3",
    "ruff>=0.11.5",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
